requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
altgraph==0.17.4
    # via pyinstaller
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.12.3
blinker==1.8.2
    # via flask
boto3==1.35.10
    # via moto
botocore==1.35.10
    # via
    #   boto3
    #   moto
    #   s3transfer
certifi==2024.8.30
    # via requests
cffi==1.17.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cryptography==43.0.0
    # via moto
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flask==3.0.3
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.24.6
    # via
    #   tokenizers
    #   transformers
idna==3.8
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via
    #   flask
    #   moto
    #   torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
moto==5.0.13
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.0
    # via transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.68
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==24.1
    # via
    #   huggingface-hub
    #   pyinstaller
    #   pyinstaller-hooks-contrib
    #   transformers
py-healthcheck==1.10.1
py-partiql-parser==0.5.5
pycparser==2.22
    # via cffi
pydantic==2.8.2
pydantic-core==2.20.1
    # via pydantic
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
    # via pyinstaller
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
python-dotenv==1.0.1
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   responses
    #   transformers
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   moto
    #   responses
    #   transformers
responses==0.25.3
    # via moto
ruff==0.6.3
s3transfer==0.10.2
    # via boto3
safetensors==0.4.4
    # via transformers
setuptools==74.0.0
    # via
    #   pyinstaller
    #   pyinstaller-hooks-contrib
six==1.16.0
    # via
    #   py-healthcheck
    #   python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sympy==1.13.2
    # via torch
tokenizers==0.19.1
    # via transformers
torch==2.4.0
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   transformers
transformers==4.44.2
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.2
    # via
    #   botocore
    #   requests
    #   responses
waitress==3.0.0
werkzeug==3.0.4
    # via
    #   flask
    #   moto
xmltodict==0.13.0
    # via moto
